feat: Add new exceptions for file management. Move exceptions to exceptions.py
