+++ pyproject.toml
@@ -44,6 +44,7 @@ mkdocs-material = "^9.1.12"
 mypy = "^1.3.0"
 pre-commit = "^3.3.1"
 pytest = "^7.3.1"
+types-pyyaml = "^6.0.12.10"
 
 [build-system]
 requires = ["poetry-core"]
